API/app/requirements.txt
Flask-RESTful
Flask
gunicorn
python-dotenv
flask-apispec
jinja2
pynamodb
boto3